+++ src/DragSelect.js
@@ -62,6 +62,11 @@ import { toArray, vect2, subscriberAliases } from './methods'
 
 class DragSelect {
   /**
+   * used to skip all current Selection and dragNdrop functionality
+   * @type {boolean}
+   */
+  continue = false
+  /**
    * @class DragSelect
    * @constructor DragSelect
    * @param {Settings} settings
@@ -183,6 +188,8 @@ class DragSelect {
       Interaction: this.Interaction,
     })
 
+    this.subscribe('Interaction:end', () => (this.continue = false))
+
     this.start()
   }
 
@@ -255,6 +262,7 @@ class DragSelect {
     this.Area.stop()
     this.Drag.stop()
     this.Selector.stop()
+    this.SelectorArea.stop()
     this.stores.KeyStore.stop()
     this.stores.PointerStore.stop()
     this.stores.ScrollStore.stop()
@@ -262,6 +270,12 @@ class DragSelect {
     if (remove) this.SelectableSet.clear()
     if (fromSelection) this.SelectedSet.clear()
   }
+  /**
+   * Utility to override DragSelect internal functionality:
+   * Break will skip the selection or dragging functionality but let everything continue to run until after the callback.
+   * Useful utility to write your own functionality/move/dragNdrop based on DragSelect pointer positions.
+   */
+  break = () => (this.continue = true)
   /**
    * Returns the current selected nodes
    * @return {DSElements}
@@ -416,18 +430,28 @@ class DragSelect {
    */
   isMultiSelect = (event) => this.stores.KeyStore.isMultiSelectKeyPressed(event)
   /**
-   * Returns the cursor position difference between start and now
+   * Utility method that returns the cursor position difference between start and now
    * @param {boolean} [usePreviousCursorDifference] if true, it will output the cursor position difference between the previous selection and now
+   * @param {boolean} [useAreaPositions] if true, it will use cursor positions relative to the area
    * @return {Vect2}
    * @deprecated
    */
-  getCursorPositionDifference(usePreviousCursorDifference = false) {
+  getCursorPositionDifference(
+    usePreviousCursorDifference = false,
+    useAreaPositions = false
+  ) {
     console.warn(
       '[DragSelect] Using .getCursorPositionDifference is deprecated. Calculate yourself instead. i.e. `.getCurrentCursorPosition().x - .getInitialCursorPosition().x`'
     )
-    const posA = this.getCurrentCursorPosition()
+    const posA = useAreaPositions
+      ? this.getCurrentCursorPositionArea()
+      : this.getCurrentCursorPosition()
     const posB = usePreviousCursorDifference
-      ? this.getPreviousCursorPosition()
+      ? useAreaPositions
+        ? this.getPreviousCursorPositionArea()
+        : this.getPreviousCursorPosition()
+      : useAreaPositions
+      ? this.getInitialCursorPositionArea()
       : this.getInitialCursorPosition()
     return vect2.calc(posA, '-', posB)
   }